Moved DefaultRequestHandler to CefSharp.Handler namespace; added comments
