--- CefSharp/DefaultRequestHandler.cs
@@ -1,96 +1,0 @@
-﻿// Copyright © 2010-2017 The CefSharp Authors. All rights reserved.
-//
-// Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
-
-using System.Security.Cryptography.X509Certificates;
-
-namespace CefSharp
-{
-    /// <summary>
-    /// Default implementation of <see cref="IRequestHandler"/>.
-    /// This class provides default implementations of the methods from <see cref="IRequestHandler"/>,
-    /// therefore providing a convenience base class for any custom request handler.
-    /// </summary>
-    public class DefaultRequestHandler : IRequestHandler
-    {
-        public virtual bool GetAuthCredentials(IWebBrowser browserControl, IBrowser browser, IFrame frame, bool isProxy, string host,
-            int port, string realm, string scheme, IAuthCallback callback)
-        {
-            callback.Dispose();
-            return false;
-        }
-
-        public virtual IResponseFilter GetResourceResponseFilter(IWebBrowser browserControl, IBrowser browser, IFrame frame,
-            IRequest request, IResponse response)
-        {
-            return null;
-        }
-
-        public virtual bool OnBeforeBrowse(IWebBrowser browserControl, IBrowser browser, IFrame frame, IRequest request, bool isRedirect)
-        {
-            return false;
-        }
-
-        public virtual CefReturnValue OnBeforeResourceLoad(IWebBrowser browserControl, IBrowser browser, IFrame frame, IRequest request,
-            IRequestCallback callback)
-        {
-            callback.Dispose();
-            return CefReturnValue.Continue;
-        }
-
-        public virtual bool OnCertificateError(IWebBrowser browserControl, IBrowser browser, CefErrorCode errorCode, string requestUrl,
-            ISslInfo sslInfo, IRequestCallback callback)
-        {
-            callback.Dispose();
-            return false;
-        }
-
-        public virtual bool OnOpenUrlFromTab(IWebBrowser browserControl, IBrowser browser, IFrame frame, string targetUrl,
-            WindowOpenDisposition targetDisposition, bool userGesture)
-        {
-            return false;
-        }
-
-        public virtual void OnPluginCrashed(IWebBrowser browserControl, IBrowser browser, string pluginPath)
-        { }
-
-        public virtual bool OnProtocolExecution(IWebBrowser browserControl, IBrowser browser, string url)
-        {
-            return false;
-        }
-
-        public virtual bool OnQuotaRequest(IWebBrowser browserControl, IBrowser browser, string originUrl, long newSize,
-            IRequestCallback callback)
-        {
-            callback.Dispose();
-            return false;
-        }
-
-        public virtual void OnRenderProcessTerminated(IWebBrowser browserControl, IBrowser browser, CefTerminationStatus status)
-        { }
-
-        public virtual void OnRenderViewReady(IWebBrowser browserControl, IBrowser browser)
-        { }
-
-        public virtual void OnResourceLoadComplete(IWebBrowser browserControl, IBrowser browser, IFrame frame, IRequest request,
-            IResponse response, UrlRequestStatus status, long receivedContentLength)
-        { }
-
-        public virtual void OnResourceRedirect(IWebBrowser browserControl, IBrowser browser, IFrame frame, IRequest request,
-            IResponse response, ref string newUrl)
-        { }
-
-        public virtual bool OnResourceResponse(IWebBrowser browserControl, IBrowser browser, IFrame frame, IRequest request,
-            IResponse response)
-        {
-            return false;
-        }
-
-        public virtual bool OnSelectClientCertificate(IWebBrowser browserControl, IBrowser browser, bool isProxy, string host, int port,
-            X509Certificate2Collection certificates, ISelectClientCertificateCallback callback)
-        {
-            callback.Dispose();
-            return false;
-        }
-    }
-}
